+++ tap/runner.py
@@ -15,6 +15,7 @@ class TAPTestResult(TextTestResult):
 
     # This attribute will store the user's desired output directory.
     OUTDIR = None
+    FORMAT = None
 
     def __init__(self, stream, descriptions, verbosity):
         super(TAPTestResult, self).__init__(stream, descriptions, verbosity)
@@ -56,6 +57,16 @@ class TAPTestResult(TextTestResult):
         return test.__class__.__name__
 
     def _description(self, test):
+        if self.FORMAT:
+            try:
+                return self.FORMAT.format(
+                    short_description=test.shortDescription(),
+                    method_name=str(test))
+            except KeyError as e:
+                exit('''Bad format string: {key}
+Replacement options are: \{short_description\} and \{method_name\}'''.format(
+                    key=e[0]))
+
         return test.shortDescription() or str(test)
 
 
@@ -72,3 +83,10 @@ class TAPTestRunner(TextTestRunner):
         """
         # Blame the lack of unittest extensibility for this hacky method.
         TAPTestResult.OUTDIR = outdir
+
+    @classmethod
+    def set_format(cls, fmt):
+        """Set the format of each line to a string
+        {method_name}: method name
+        {short_description}: short description"""
+        TAPTestResult.FORMAT = fmt